Enable the mouse
Secretly, Vim is able to understand mouse clicks, scrolling,
highlighting, and just about everything you'd be familiar with in an IDE
or GUI text editor. While there are times when using the keyboard is the
absolute fastest way to get something done, there are plent of other
times where being able to scroll through Vim or use the mouse is pretty
handy.