Set up Pathogen boilerplate
We're going to be using the Vim plugin manager called Vundle.

NOTE

For this talk, I'm assuming that you're starting from a blank slate. If
you have a ~/.vim folder, move it out of the way:

    mv ~/.vim ~/.vim.old

INSTALLATION INSTRUCIONS

Before making these changes, you'll need to install Vundle. Run this
command to install it:

    git clone https://github.com/gmarik/Vundle.vim
~/.vim/bundle/Vundle.vim

This installs Vundle.vim to a subdirectory of your ~/.vim folder. We'll
be using Vundle in upcoming steps to add lots of plugins!

The changes we made in this step are outlined right on the front page of
the Vundle project page on GitHub. They're just stock changes that need
to be made for Vundle to work.