Plugins vim-easytags and tagbar
There's a tad bit of external setup required here:

- Using your package manager, install the program `ctags`.

------------------------------------------------------------------------

There are a number of plugins being installed here.

- vim-misc is a dependency of vim-easytags

- vim-easytags is a plugin that generates tags files, which is a
  compiled index of all the functions, variables, and identifies that
  you use in your project.

- tagbar is a plugin that reads the output from ctags and displays the
  information in an accessible way inside Vim.

Most of everything is configured out of the box, and you can press ?
inside the tagbar buffer to learn how to use it (after opening it with
\b). For more information,

- http://majutsushi.github.io/tagbar/
- https://github.com/xolox/vim-easytags
- https://github.com/xolox/vim-misc

These give you the power to see what kinds of methods, variables,
functions, and other types of declarations you have in your files.

If in the future you're wondering why `tags` files start showing up in
random places, this step is your answer.