Plugin ctrlp
ctrlp is a magical Vim plugin for finding files quickly. It lets you
"fuzzy find" files, which you'll be familiar with if you use Sublime a
lot. Basically, you can type a subset of the letters in a file's name,
and the plugin will be smart enough to figure it out.

- From within normal mode, press Ctrl + P, and start typing.
    - You'll see that a list of all files in this subdirectory are
      getting filtered as you type,
    - you only have to type a few characters to bring up the file that
      you need.
- Press Enter to open the file, or Ctrl + T to open in a new tab.