Plugin vim-superman
This plugin actually allows you to view man pages in Vim! Just add the
text

    vman() {
      vim -c "SuperMan $*"

      if [ "$?" != "0" ]; then
        echo "No manual entry for $*"
      fi
    }

to your ~/.bashrc, ~/.zshrc, or other config file, and then run

    vman git

to open the man page for git in Vim!

You can also press K on a word in Vim to bring up the man page for that
word.