Syntax plugins
While Vim has a ton of built-in support for different languages'
syntaxes, sometimes for cutting edge stuff it falls behind. In these
cases, a quick search will usually turn up a syntax highlighting plugin
for the file you need!

Here we've included a few examples of this. Feel free to delete the ones
you don't need and search for more to add that you do.